Fix: textarea text cursor cannot be moved by clicking
The text area can be focused when clicked but the text cursor cannot
be moved to specified position by clicking. It can only be moved by
using arrow keys on the keyboard. This is likely due to the parent
todoItem div's draggable="true" hijacking mouse event on the textarea to
prepare for drag events.